requirements.txt
# ============================================================
# CryptoStream AI — Python Dependencies
# Shared across: producer.py, lake_writer.py, load_tester.py,
#                flink_processor.py, mcp_server/, test_e2e.py
# ============================================================

# ── Kafka ─────────────────────────────────────────────────
kafka-python          # Kafka producer/consumer for ingestion & lake writer
websocket-client      # Binance WebSocket stream (producer.py)

# ── Database ───────────────────────────────────────────────
psycopg2-binary       # PostgreSQL client (MCP server, test_e2e.py, Airflow DAGs)
sqlalchemy            # ORM / connection pooling for MCP server

# ── Data Lake ─────────────────────────────────────────────
pandas                # DataFrame transformations in lake_writer.py
pyarrow               # Parquet read/write with explicit schema enforcement

# ── Utilities ─────────────────────────────────────────────
python-dotenv         # Load .env files for local dev secret management

# ── MCP Server (Phase 6 — AI Agent Integration) ──────────
fastapi               # REST API framework for MCP server
uvicorn               # ASGI server for FastAPI
requests              # HTTP client for health checks and MCP tests
# ── AI / LLM ──────────────────────────────────────────────
google-genai          # Google's next-gen AI SDK for Gemini 2.5 Flash
python-dotenv         # Load .env files for local dev secret management
slowapi               # Rate limiting middleware for FastAPI (API key protection)
pydantic              # Data validation for MCP request/response models
# ── Real-time Analytics (Phase 2 Additions) ─────────
aiokafka              # Async Kafka client for FastAPI WebSocket bridge
lightweight-charts    # Financial charts for institutional deep-dives

# ── PyFlink (uncomment when running inside Flink container) ──
# apache-flink==1.18.1
# apache-flink-libraries==1.18.1
